feat: added symbolizer
